+++ docs/news.rst
@@ -6,6 +6,7 @@ v4.3 (unreleased)
 
 Lab
 ^^^
+* If there is only a single HTML table, show it when the page loads.
 * Remove broken ``--log-level`` command line parameter. You can use
   ``tools.configure_logging(logging.DEBUG)`` to enable debug messages instead.